move version # outside OSM module
